Finally after a lot of pain in the ass, jwt setup is ready. Login is permitted without authentication for getting token. Then all book api requests get filtered and authenticated using jwt tokens.

+++ src/main/resources/db/migration/V1__BOOKS_SCHEMA.sql
@@ -1,7 +1,8 @@
 CREATE TABLE books(
-    id NUMERIC(19,0) NOT NULL PRIMARY KEY,
+    id INT NOT NULL PRIMARY KEY,
     title VARCHAR NOT NULL,
     description VARCHAR NOT NULL,
     author VARCHAR NOT NULL,
     release_year INT NOT NULL
 )
+-- id NUMERIC(19,0) NOT NULL PRIMARY KEY,
